#136 move account related stuff to the models package
